+++ src/invoices/CreateInvoiceLine/index.js
@@ -1,0 +1,1 @@
+export { default as CreateInvoiceLine } from './CreateInvoiceLine';