+++ src/invoices/InvoiceLineForm/index.js
@@ -1,1 +1,2 @@
-export { default } from './InvoiceLineForm';
+export { default as InvoiceLineForm } from './InvoiceLineForm';
+export { default as InvoiceLineFormContainer } from './InvoiceLineFormContainer';