+++ src/invoices/InvoicesList/InvoicesListFilters/index.js
@@ -1,0 +1,2 @@
+export { default as InvoicesListFilters } from './InvoicesListFilters';
+export { default as InvoicesListFiltersContainer } from './InvoicesListFiltersContainer';