+++ src/invoices/InvoicesList/index.js
@@ -1,1 +1,2 @@
-export { default } from './InvoicesList';
+export { default as InvoicesList } from './InvoicesList';
+export { default as InvoicesListContainer } from './InvoicesListContainer';